requirements.txt
Flask
opencv-python
firebase-admin
numpy
matplotlib
Pillow